+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-16 12:09:28
-// Commit: feat(security): add bi-directional event bus handler for security
+// Updated at: 2026-05-17 15:14:08
+// Commit: fix(security): fix memory leak by cancelling security subscription
 
 class securityComponentState {
-  final String commitRef = '1203';
-  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
-  final DateTime timestamp = DateTime.parse('2026-05-16T12:09:28+05:00');
+  final String commitRef = '1224';
+  final String lastMessage = 'fix(security): fix memory leak by cancelling security subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-17T15:14:08+05:00');
 }